# docs/swagger.yaml
basePath: /
definitions:
  model.DataFlowDiagram:
    properties:
      id:
        description: A randomised string e.g. dfd_abcdef123456
        type: string
      root:
        type: string
    type: object
  model.Threat:
    properties:
      componentID:
        description: |-
          Where in the model this threat occurs. A FlowID, a ComponentID or
          a ComponentGroupID
        type: string
      controls:
        items:
          type: string
        type: array
      description:
        minLength: 2
        type: string
      id:
        description: A randomised string e.g. th_abcdef123456
        type: string
      inherentRisk:
        type: string
      stride:
        minLength: 2
        type: string
    type: object
  model.ThreatModel:
    properties:
      dfd:
        $ref: '#/definitions/model.DataFlowDiagram'
      id:
        description: A randomised string e.g. th_abcdef123456
        type: string
      threats:
        items:
          $ref: '#/definitions/model.Threat'
        type: array
      title:
        minLength: 2
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@threatplane.io
    name: ThreatPlane
    url: http://www.threatplane.io
  description: The API used to interact with Threat Models
  license:
    name: Commercial licence
    url: https://threatplane.io/terms
  title: ThreatPlane Threat Model API
  version: "1.0"
paths:
  /api/v1/{threatModelID}:
    get:
      operationId: get-threat-model-by-id
      parameters:
      - description: threat model ID
        in: path
        name: threatModelID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.ThreatModel'
        "404":
          description: Not Found
          schema:
            type: string
      summary: Retrieves a threatModel by ID.
  /api/v1/threatModel:
    get:
      operationId: get-all-threat-models
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.ThreatModel'
            type: array
      summary: Get all Threat Models
    patch:
      operationId: update-threat-model
      parameters:
      - description: todo data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/model.ThreatModel'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.ThreatModel'
        "400":
          description: Nope
          schema:
            type: string
      summary: Update a new Threat Model
    put:
      operationId: create-threat-model
      parameters:
      - description: todo data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/model.ThreatModel'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.ThreatModel'
        "400":
          description: Nope
          schema:
            type: string
      summary: Create a new Threat Model
swagger: "2.0"